Add EVAL_SKIP_ERRORS environment variable and track instances that reach maximum retries
