+++ evaluation/benchmarks/swe_bench/run_infer.py
@@ -27,6 +27,7 @@ from evaluation.utils.shared import (
     EvalMetadata,
     EvalOutput,
     assert_and_raise,
+    check_maximum_retries_exceeded,
     codeact_user_response,
     get_default_sandbox_config_for_eval,
     get_metrics,
@@ -60,26 +61,9 @@ from openhands.runtime.base import Runtime
 from openhands.utils.async_utils import call_async_from_sync
 from openhands.utils.shutdown_listener import sleep_if_should_continue
 
-
-def check_maximum_retries_exceeded(eval_output_dir):
-    """Check if maximum_retries_exceeded.txt exists and output a message."""
-    retries_file_path = os.path.join(eval_output_dir, 'maximum_retries_exceeded.txt')
-    if os.path.exists(retries_file_path):
-        with open(retries_file_path, 'r') as f:
-            content = f.read().strip()
-            if content:
-                failed_instances = content.split('\n')
-                failed_count = len(failed_instances)
-                logger.info(
-                    f'ATTENTION: {failed_count} instances reached maximum error retries and were skipped.'
-                )
-                logger.info(f'These instances are listed in: {retries_file_path}')
-
-
 USE_HINT_TEXT = os.environ.get('USE_HINT_TEXT', 'false').lower() == 'true'
 RUN_WITH_BROWSING = os.environ.get('RUN_WITH_BROWSING', 'false').lower() == 'true'
-BenchMode = Literal['swe', 'swt', 'swt-ci']
-RUN_WITH_BROWSING = os.environ.get('RUN_WITH_BROWSING', 'false').lower() == 'true'
+EVAL_SKIP_ERRORS = os.environ.get('EVAL_SKIP_ERRORS', 'false').lower() == 'true'
 BenchMode = Literal['swe', 'swt', 'swt-ci']
 
 
@@ -879,7 +863,7 @@ if __name__ == '__main__':
             max_retries=5,
         )
 
-        # Check if maximum_retries_exceeded.txt exists and output a message
+        # Check if any instances reached maximum retries
         check_maximum_retries_exceeded(metadata.eval_output_dir)
     else:
         critic = AgentFinishedCritic()
@@ -930,6 +914,9 @@ if __name__ == '__main__':
                 * 60,  # 8 hour PER instance should be more than enough
                 max_retries=5,
             )
+
+            # Check if any instances reached maximum retries
+            check_maximum_retries_exceeded(metadata.eval_output_dir)
 
             # When eval is done, we update eval_ids to the instances that failed the current attempt
             instances_failed = []
